Create README.md
<b>What was your motivation?</b>
Our motivation for building this part comes from the customers request to not have information stored in the cloud by the big corperations. And this project became the proof of concept build. It is the first working version to prove that the concept could work. 

------Why did you build this project?------
This API is a part of the whole application for a decentralized social media platform.

------What problem does it solve?------
 It was made to communicate between nodes in the network, so that all information stays local.

------What did you learn?------
The lessons learnt in this project includes.
 
Communication between frontend and backend.
How to use a mailservice.
How to create a standard CRUD API
How to send information from the Database (MariaDB)
How to encrypt text.

------What makes your project stand out? If your project has a lot of features, consider adding a "Features" section and listing them here.------

FEATURES
Using mail for communication between users.

DOCUMENTATION